Logged in User Update and Delete Contact

+++ controllers/contactController.js
@@ -49,6 +49,10 @@ const updateContact = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error("Contact not found");
     }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User does not have permission to access other user contacts")
+    }
 
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
@@ -67,6 +71,12 @@ const deleteContact = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error("Contact not found");
     }
+
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User does not have permission to access other user contacts")
+    }
+
     await contact.deleteOne();  
     res.status(200).json(contact)
 })
